Convert readme to markdown format (#2)
* convert readme to markdown

Just to get it organized the same as the other cores.

Minor adjustments to formatting, added currently supported games for now with links so people can see the history. Linked to Universal as a brand so it can help inform people a little more. Linked to github account of Mike Coates. Linked to MiSTer and MiSTer wiki a couple times like all the others. Added standard link to the hardware in the title. Modified grammar very slightly, not trying to rewrite your readme entirely as it's already quite good. ;)

* Slight cleanup, found a couple things